wrap functions up in a class
also renaming db -> load

all to more closely mirror the mds.api interface and hopefully make
using this library easier
