+++ mds/db/__init__.py
@@ -3,5 +3,5 @@ Work with MDS Provider data.
 """
 
 from mds.db import db
-from mds.db.db import data_engine, load_status_changes, load_trips
+from mds.db.load import data_engine, ProviderDataLoader
